PyTorch_BrainImage_AWS_Lambda/requirements.txt
##torchvision==0.15.2 
#torch==2.0.1
#numpy==1.24.3
#nltk==3.8.1
serverless_wsgi==3.0.2
pillow==9.4.0
Werkzeug==2.3
flask
torchvision==0.15.2 
torch==2.0.1

